Refactor WebGLCanvas import and shader constants
Moved WebGLCanvas to a top-level src directory and updated all imports accordingly. Refactored GLSL shaders to use named constants for improved readability and maintainability. Unified disposable management in shader managers by using a shared 'disposables' set instead of per-class private sets.
