+++ templates/shader/src/WebGLCanvas.tsx
@@ -1,13 +1,12 @@
 import { forwardRef } from 'react'
 import { Atom, useValue } from 'tldraw'
-import { WebGLManagerConfig } from '../WebGLManager'
+import { WebGLManagerConfig } from './WebGLManager'
 
 export const WebGLCanvas = forwardRef<
 	HTMLCanvasElement,
 	{ config: Atom<WebGLManagerConfig, unknown> }
 >(({ config }, ref) => {
 	const isPixelated = useValue('config', () => config.get().pixelate, [])
-	console.log('isPixelated', isPixelated)
 	const canvasClassName = `shader-canvas${isPixelated ? ' shader-canvas-pixelated' : ''}`
 	return <canvas ref={ref} className={canvasClassName} />
 })
