+++ templates/shader/src/rainbow/RainbowShaderManager.ts
@@ -27,8 +27,6 @@ export class RainbowShaderManager extends WebGLManager<ShaderManagerConfig> {
 	private geometries: Geometry[] = []
 	private maxSegments: number = 2000
 
-	private _disposables = new Set<() => void>()
-
 	constructor(
 		editor: Editor,
 		canvas: HTMLCanvasElement,
@@ -149,7 +147,7 @@ export class RainbowShaderManager extends WebGLManager<ShaderManagerConfig> {
 			gl2.bindVertexArray(null)
 		}
 
-		this._disposables.add(react('dependencies', this.tick))
+		this.disposables.add(react('dependencies', this.tick))
 
 		this.tick()
 	}
@@ -247,9 +245,6 @@ export class RainbowShaderManager extends WebGLManager<ShaderManagerConfig> {
 	}
 
 	onDispose = (): void => {
-		this._disposables.forEach((dispose) => dispose())
-		this._disposables.clear()
-
 		if (this.gl) {
 			if (this.vao && this.gl instanceof WebGL2RenderingContext) {
 				this.gl.deleteVertexArray(this.vao)
