+++ templates/shader/src/shadow/fragment.glsl
@@ -12,11 +12,14 @@ uniform float u_zoom;
 uniform vec2 u_lightPos;
 uniform float u_shadowContrast;
 
-#define LIGHT_SHADOW_ALPHA 0.5
-#define DARK_SHADOW_ALPHA 0.15
+// Constants
+#define MAX_SEGMENTS 512
+const float LIGHT_SHADOW_ALPHA = 0.5;
+const float DARK_SHADOW_ALPHA = 0.15;
+const float EPSILON_SMALL = 0.001;
+const float DARK_MODE_THRESHOLD = 0.5;
 
 // Geometry data
-#define MAX_SEGMENTS 512
 uniform vec4 u_segments[MAX_SEGMENTS]; // xy = start, zw = end
 uniform float u_segmentCount;
 
@@ -26,17 +29,17 @@ float raySegmentIntersect(vec2 rayOrigin, vec2 rayDir, vec2 segStart, vec2 segEn
 	vec2 v1 = rayOrigin - segStart;
 	vec2 v2 = segEnd - segStart;
 	vec2 v3 = vec2(-rayDir.y, rayDir.x);
-	
+
 	float det = dot(v2, v3);
-	if (abs(det) < 0.001) return -1.0;
-	
+	if (abs(det) < EPSILON_SMALL) return -1.0;
+
 	float t1 = (v2.x * v1.y - v2.y * v1.x) / det;
 	float t2 = dot(v1, v3) / det;
-	
+
 	if (t1 >= 0.0 && t2 >= 0.0 && t2 <= 1.0) {
 		return t1;
 	}
-	
+
 	return -1.0;
 }
 
@@ -58,7 +61,7 @@ bool isVisible(vec2 point, vec2 lightPos) {
 
 		// If intersection exists and is closer than the light, we're blocked
 		// Using small epsilon for normalized coordinates (0-1 range)
-		if (t > 0.001 && t < distToLight - 0.001) {
+		if (t > EPSILON_SMALL && t < distToLight - EPSILON_SMALL) {
 			return false;
 		}
 	}
@@ -72,7 +75,7 @@ void main() {
 	// Check if this pixel is visible from the light
 	bool visible = isVisible(v_uv, u_lightPos);
 
-	bool isDark = u_darkMode > 0.5;
+	bool isDark = u_darkMode > DARK_MODE_THRESHOLD;
 
 	if (isDark) {
 		if (visible) {
